Fixed a bug in current deposit and removed unnecessary files

+++ include/data/typedefs.h
@@ -15,7 +15,7 @@ typedef double Scalar;
 typedef std::bitset<3> Stagger_t;
 
 typedef double Mom_t;
-typedef float Pos_t;
+typedef double Pos_t;
 
 typedef std::size_t Index_t;
